requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
aiogram=2.13=pypi_0
aiohttp=3.7.4.post0=pypi_0
apscheduler=3.7.0=pypi_0
asgiref=3.3.4=pypi_0
async-timeout=3.0.1=pypi_0
attrs=21.2.0=pypi_0
babel=2.9.1=pypi_0
blas=1.0=mkl
bzip2=1.0.8=h7b6447c_0
ca-certificates=2021.5.25=h06a4308_1
certifi=2021.5.30=py38h06a4308_0
chardet=4.0.0=pypi_0
click=8.0.1=pypi_0
cudatoolkit=11.1.74=h6bb024c_0
django=3.2.4=pypi_0
ffmpeg=4.3=hf484d3e_0
flask=2.0.0=pypi_0
flask-pymongo=2.3.0=pypi_0
freetype=2.10.4=h5ab3b9f_0
gmp=6.2.1=h2531618_2
gnutls=3.6.15=he1e5248_0
idna=2.10=pypi_0
intel-openmp=2021.2.0=h06a4308_610
itsdangerous=2.0.1=pypi_0
jinja2=3.0.1=pypi_0
jpeg=9b=h024ee3a_2
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.33.1=h53a641e_7
libffi=3.3=he6710b0_2
libgcc-ng=9.1.0=hdf63c60_0
libiconv=1.15=h63c8f33_5
libidn2=2.3.1=h27cfd23_0
libpng=1.6.37=hbc83047_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtasn1=4.16.0=h27cfd23_0
libtiff=4.2.0=h85742a9_0
libunistring=0.9.10=h27cfd23_0
libuv=1.40.0=h7b6447c_0
libwebp-base=1.2.0=h27cfd23_0
lz4-c=1.9.3=h2531618_0
markupsafe=2.0.1=pypi_0
mkl=2021.2.0=h06a4308_296
mkl-service=2.3.0=py38h27cfd23_1
mkl_fft=1.3.0=py38h42c9631_2
mkl_random=1.2.1=py38ha9443f7_2
multidict=5.1.0=pypi_0
ncurses=6.2=he6710b0_1
nettle=3.7.2=hbbd107a_1
ninja=1.10.2=hff7bd54_1
numpy=1.20.2=py38h2d18471_0
numpy-base=1.20.2=py38hfae3a4d_0
olefile=0.46=py_0
openh264=2.1.0=hd408876_0
openssl=1.1.1k=h27cfd23_0
pillow=8.2.0=py38he98fc37_0
pip=21.0.1=py38h06a4308_0
pymongo=3.11.4=pypi_0
python=3.8.10=hdb3f193_7
pytorch=1.8.1=py3.8_cuda11.1_cudnn8.0.5_0
pytz=2021.1=pypi_0
readline=8.1=h27cfd23_0
requests=2.25.1=pypi_0
setuptools=52.0.0=py38h06a4308_0
six=1.16.0=pypi_0
sqlite=3.35.4=hdfb4753_0
sqlparse=0.4.1=pypi_0
tk=8.6.10=hbc83047_0
torchaudio=0.8.1=py38
torchvision=0.9.1=py38_cu111
typing-extensions=3.10.0.0=pypi_0
typing_extensions=3.7.4.3=pyha847dfd_0
tzlocal=2.1=pypi_0
urllib3=1.26.4=pypi_0
werkzeug=2.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
whitenoise=5.2.0=pypi_0
xz=5.2.5=h7b6447c_0
yarl=1.6.3=pypi_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.9=haebb681_0
